basereader: move spatial and temporal domain into ReaderDomain
